warning fix
git-svn-id: https://swig.svn.sourceforge.net/svnroot/swig/trunk@12827 626c5289-ae23-0410-ae9c-e8d60b6d4f22

+++ Source/Swig/misc.c
@@ -1165,7 +1165,7 @@ String *replace_captures(const char *input, String *subst, int captures[])
     /* Handle substitution */
     if (*p == '\0') {
       Putc('\\', result);
-    } else if (isdigit(*p)) {
+    } else if (isdigit((int)*p)) {
       int group = *p++ - '0';
       int l = captures[group*2], r = captures[group*2 + 1];
       if (l != -1) {
